use size_t instead of unsigned long long
There's no point in using a data type which can represent lengths beyond
the total size of the address space on the target platform.

Review: http://cr.bittorrent.com/4661265418944512

+++ src/DhtImpl.h
@@ -207,7 +207,7 @@ public:
 
 	// It's critical that this overload is distinct from the format string
 	// overload. Otherwise they are too similar and error prone
-	smart_buffer& operator() (int64 len, unsigned char const* value) {
+	smart_buffer& operator() (size_t len, unsigned char const* value) {
 		assert(buffer + len < end);
 		if (buffer + len >= end) return *this;
